refactor(gradle-plugin): Move Gradle model extensions to a separate file
These functions do not need class context, and this nicely resembles the
existing `MavenModelextensions`.
